+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/presentation/v1/ReviewController.java
@@ -61,7 +61,7 @@ class ReviewController implements ReviewApi {
 	@Override
 	@GetMapping("/shop-admin/{shopAdminId}/await")
 	public ResponseEntity<SuccessResponse<PageResponse<GetAwaitReviewResponse>>> getAwaitReviewsByShopAdmin(
-		@PathVariable("shopAdminId") long shopAdminId,
+		@PathVariable("shopAdminId") Integer shopAdminId,
 		@RequestParam(value = "size", required = false, defaultValue = "10") int size,
 		@RequestParam(value = "page", required = false, defaultValue = "0") int page,
 		@RequestParam(name = "direction", required = false, defaultValue = "DESC")
